requirements_pip.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
imageio==2.35.1
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
qimage2ndarray==1.10.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
